Prevent stack overflow when missing href
When a resource doesn't have a `href` attribute, it hits method_missing,
which tries to load the full resource, which subsequently calls href,
and then we enter a loop and ultimately the stack overflows.

The solution is to add href to attr_accessor so that calls to href never
hit method_missing.